refactor(Button): remove the interface to index file
