--- src/components/inputs/Button/interfaces/Button.ActionStyle.interface.ts
@@ -1,4 +1,0 @@
-export interface IActionStyle {
-  filled: string;
-  stroke: string;
-}
