--- src/components/inputs/Button/interfaces/Button.Actions.interface.ts
@@ -1,11 +1,0 @@
-import { IActionStyle } from "./Button.ActionStyle.interface";
-
-export interface IActions {
-  primary: IActionStyle;
-  secondary: IActionStyle;
-  confirm: IActionStyle;
-  warning: IActionStyle;
-  remove: IActionStyle;
-  help: IActionStyle;
-  disabled: IActionStyle;
-}
