--- src/components/inputs/Button/interfaces/Button.BackgroundColorVariant.interface.ts
@@ -1,6 +1,0 @@
-import { IActions } from "./Button.Actions.interface";
-
-export interface IBackgroundColorVariant {
-  normal: Partial<IActions> | any;
-  hover: Partial<IActions> | any;
-}
