--- src/components/inputs/Button/interfaces/Button.BorderColorVariant.interface.ts
@@ -1,6 +1,0 @@
-import { IActions } from "./Button.Actions.interface";
-
-export interface IBorderColorVariant {
-  normal: Partial<IActions> | any;
-  hover: Partial<IActions> | any;
-}
