--- src/components/inputs/Button/interfaces/Button.BorderColors.interface.ts
@@ -1,8 +1,0 @@
-import { IBorderColorVariant } from "./Button.BorderColorVariant.interface";
-import { IVariantStyle } from "./Button.VariantStyle.interface";
-
-export interface IBorderColors {
-  filled: IVariantStyle | string;
-  outlined: IVariantStyle | IBorderColorVariant;
-  none: IVariantStyle | string;
-}
