--- src/components/inputs/Button/interfaces/Button.ColorPalette.interface.ts
@@ -1,9 +1,0 @@
-export interface IColorPalette {
-  primary: string;
-  secondary: string;
-  confirm: string;
-  warning: string;
-  remove: string;
-  help: string;
-  disabled?: string;
-}
