--- src/components/inputs/Button/interfaces/Button.Cursors.interface.ts
@@ -1,5 +1,0 @@
-export interface ICursors {
-  pointer: string;
-  notAllowed: string;
-  progress: string;
-}
