--- src/components/inputs/Button/interfaces/Button.HoverColors.interface.ts
@@ -1,8 +1,0 @@
-export interface IHoverColors {
-  primary: string;
-  secondary: string;
-  confirm: string;
-  warning: string;
-  remove: string;
-  help: string;
-}
