--- src/components/inputs/Button/interfaces/Button.Spacing.interface.ts
@@ -1,6 +1,0 @@
-import { ISpacingDetail } from "./Button.SpacingDetail.interface";
-
-export interface ISpacing {
-  compact: ISpacingDetail;
-  wide: ISpacingDetail;
-}
