--- src/components/inputs/Button/interfaces/Button.SpacingDetail.interface.ts
@@ -1,4 +1,0 @@
-export interface ISpacingDetail {
-  height: string;
-  minWidth: string;
-}
