--- src/components/inputs/Button/interfaces/Button.TextColors.interface.ts
@@ -1,7 +1,0 @@
-import { IVariantStyle } from "./Button.VariantStyle.interface";
-
-export interface ITextColors {
-  filled: IVariantStyle;
-  outlined: IVariantStyle;
-  none: IVariantStyle;
-}
