--- src/components/inputs/Button/interfaces/Button.VariantStyle.interface.ts
@@ -1,6 +1,0 @@
-import { IColorPalette } from "./Button.ColorPalette.interface";
-
-export interface IVariantStyle {
-  normal: IColorPalette;
-  hover: Partial<IColorPalette> | any;
-}
