--- src/components/inputs/Button/interfaces/Button.interface.ts
@@ -1,20 +1,0 @@
-import { Appearance } from "../types/Button.Appearances.type";
-import { Type } from "../types/Button.Types.type";
-import { Spacing } from "../types/Button.Spacings.type";
-import { Variant } from "../types/Button.Variants.type";
-
-export interface IButtonProps {
-  children: React.ReactNode;
-  appearance?: Appearance;
-  isLoading?: boolean;
-  isDisabled?: boolean;
-  isdisabled?: number;
-  iconBefore?: React.ReactElement;
-  iconAfter?: React.ReactElement;
-  type?: Type;
-  spacing?: Spacing;
-  variant?: Variant;
-  isFullWidth?: boolean;
-  handleClick?: (e?: Event) => void;
-  path?: string;
-}
