--- src/components/inputs/Button/interfaces/button.BackgroundColors.interface.ts
@@ -1,7 +1,0 @@
-import { IBackgroundColorVariant } from "./Button.BackgroundColorVariant.interface";
-
-export interface IBackgroundColors {
-  filled: IBackgroundColorVariant;
-  outlined: string;
-  none: string;
-}
