+++ src/components/inputs/Button/stories/Button.Variants.stories.tsx
@@ -1,7 +1,8 @@
 import { BrowserRouter } from "react-router-dom";
 import { MdAdd } from "react-icons/md";
 
-import { Button } from "../index";
+import { variants } from "../types/Button.Variants.type";
+import { Button, IButtonProps } from "..";
 import { StyledFlex } from "./stories.styles";
 
 import {
@@ -18,8 +19,6 @@ import {
   handleClick,
   path,
 } from "./props";
-import { IButtonProps } from "../interfaces/Button.interface";
-import { variants } from "../types/Button.Variants.type";
 
 const story = {
   title: "inputs/Button",
